requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Reinforcement Learning
stable-baselines3>=2.0.0
gymnasium>=0.28.0
tensorboard>=2.13.0

# Deep learning utilities
accelerate>=0.24.0
datasets>=2.14.0
tokenizers>=0.14.0
safetensors>=0.3.0

# Configuration and logging
pyyaml>=6.0
wandb>=0.15.0
rich>=13.0.0
tqdm>=4.65.0

# Scientific computing
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Model optimization
bitsandbytes>=0.41.0
flash-attn>=2.3.0

# Development and testing
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: For distributed training
# deepspeed>=0.10.0
# fairscale>=0.4.0